Add type annotations checked via mypy

+++ util/discord.py
@@ -7,12 +7,14 @@ import discord
 import discord.abc
 import string
 import logging
+from typing import (Any, List, Callable, Iterable, Optional, Union, Awaitable,
+    TypeVar, Protocol, AsyncContextManager)
 import discord_client
 import plugins
 
-logger = logging.getLogger(__name__)
+logger: logging.Logger = logging.getLogger(__name__)
 
-def unsafe_hook_event(name, fun):
+def unsafe_hook_event(name: str, fun: Callable[..., Awaitable[None]]) -> None:
     if not asyncio.iscoroutinefunction(fun):
         raise TypeError("expected coroutine function")
 
@@ -22,7 +24,8 @@ def unsafe_hook_event(name, fun):
     if not hasattr(client, method_name):
         # Insert a hook that is actually a bound method of "of" a list. The list
         # contains hooks that are to be executed.
-        async def event_hook(hooks, *args, **kwargs):
+        async def event_hook(hooks: List[Callable[..., Awaitable[None]]],
+            *args: Any, **kwargs: Any) -> None:
             for hook in list(hooks):
                 try:
                     await hook(*args, **kwargs)
@@ -31,17 +34,19 @@ def unsafe_hook_event(name, fun):
                         "Exception in {} hook {}".format(method_name, hook),
                         exc_info=True)
         event_hook.__name__ = method_name
-        client.event(event_hook.__get__([]))
+        client.event(event_hook.__get__([])) # type: ignore
 
     getattr(client, method_name).__self__.append(fun)
 
-def unsafe_unhook_event(name, fun):
+def unsafe_unhook_event(name: str, fun: Callable[..., Awaitable[None]]) -> None:
     method_name = "on_" + name
     client = discord_client.client
     if hasattr(client, method_name):
         getattr(client, method_name).__self__.remove(fun)
 
-def event(name):
+def event(name: str) -> Callable[
+        [Callable[..., Awaitable[None]]],
+        Callable[..., Awaitable[None]]]:
     """
     discord.py doesn't allow multiple functions to register for the same event.
     This decorator fixes that. Takes the event name without "on_" Example usage:
@@ -52,32 +57,36 @@ def event(name):
     This function registers a finalizer that removes the registered function,
     and hence should only be called during plugin initialization.
     """
-    def decorator(fun):
+    def decorator(fun: Callable[..., Awaitable[None]]
+        ) -> Callable[..., Awaitable[None]]:
         unsafe_hook_event(name, fun)
         @plugins.finalizer
-        def finalizer():
+        def finalizer() -> None:
             unsafe_unhook_event(name, fun)
         return fun
     return decorator
 
 class CodeBlock:
     __slots__ = "text", "language"
+    text: str
+    language: Optional[str]
 
-    def __init__(self, text, language=None):
+    def __init__(self, text: str, language: Optional[str] = None):
         self.text = text
         self.language = language
 
-    def __str__(self):
+    def __str__(self) -> str:
         text = self.text.replace("``", "`\u200D`")
         return "```{}\n".format(self.language or "") + text + "```"
 
 class Inline:
     __slots__ = "text"
+    text: str
 
-    def __init__(self, text):
+    def __init__(self, text: str):
         self.text = text
 
-    def __str__(self):
+    def __str__(self) -> str:
         text = self.text
         if "`" in text:
             if "``" in text:
@@ -104,7 +113,7 @@ class Formatter(string.Formatter):
 
     __slots__ = ()
 
-    def convert_field(self, value, conversion):
+    def convert_field(self, value: Any, conversion: str) -> Any:
         if conversion == "i":
             return str(Inline(str(value)))
         elif conversion == "b":
@@ -122,7 +131,7 @@ class Formatter(string.Formatter):
             elif isinstance(value, int):
                 return "<@&{}>".format(value)
         elif conversion == "c":
-            if isinstance(value, discord.abc.Messageable):
+            if isinstance(value, discord.TextChannel):
                 return "<#{}>".format(value.id)
             elif isinstance(value, discord.CategoryChannel):
                 return "<#{}>".format(value.id)
@@ -130,38 +139,50 @@ class Formatter(string.Formatter):
                 return "<#{}>".format(value)
         return super().convert_field(value, conversion)
 
-    def format_field(self, value, fmt):
+    def format_field(self, value: Any, fmt: str) -> Any:
         if isinstance(value, CodeBlock):
             if fmt:
                 value.language = fmt
             return str(value)
         return super().format_field(value, fmt)
 
-formatter = Formatter()
+formatter: string.Formatter = Formatter()
 format = formatter.format
 
 class UserError(Exception):
     __slots__ = "text"
 
-    def __init__(self, text, *args, **kwargs):
+    def __init__(self, text: str, *args: Any, **kwargs: Any):
         if args or kwargs:
             text = format(text, *args, **kwargs)
         super().__init__(text)
         self.text = text
 
-def smart_find(name_or_id, iterable):
+class NamedType(Protocol):
+    id: int
+    name: str
+
+class NicknamedType(Protocol):
+    id: int
+    name: str
+    nick: str
+
+T = TypeVar("T", bound=Union[NamedType, NicknamedType])
+
+def smart_find(name_or_id: str, iterable: Iterable[T]) -> Optional[T]:
     """
     Find an object by its name or id. We try an exact id match, then the
     shortest prefix match, if unique among prefix matches of that length, then
     an infix match, if unique.
     """
+    int_id: Optional[int]
     try:
         int_id = int(name_or_id)
     except ValueError:
         int_id = None
-    prefix_match = None
-    prefix_matches = []
-    infix_matches = []
+    prefix_match: Optional[T] = None
+    prefix_matches: List[str] = []
+    infix_matches: List[T] = []
     for x in iterable:
         if x.id == int_id:
             return x
@@ -170,57 +191,55 @@ def smart_find(name_or_id, iterable):
                 prefix_matches = []
             prefix_matches.append(x.name)
             prefix_match = x
-        elif getattr(x, "nick", None) != None and x.nick.startswith(name_or_id):
-            if prefix_matches and len(x.nick) < len(prefix_matches[0]):
-                prefix_matches = []
-            prefix_matches.append(x.nick)
-            prefix_match = x
-        elif name_or_id in x.name:
-            infix_matches.append(x)
-        elif getattr(x, "nick", None) != None and name_or_id in x.nick:
-            infix_matches.append(x)
+        else:
+            nick = getattr(x, "nick", None)
+            if nick is not None and nick.startswith(name_or_id):
+                if prefix_matches and len(nick) < len(prefix_matches[0]):
+                    prefix_matches = []
+                prefix_matches.append(nick)
+                prefix_match = x
+            elif name_or_id in x.name:
+                infix_matches.append(x)
+            elif nick is not None and name_or_id in nick:
+                infix_matches.append(x)
     if len(prefix_matches) == 1:
         return prefix_match
     if len(infix_matches) == 1:
         return infix_matches[0]
     return None
 
-class TempMessage:
+class TempMessage(AsyncContextManager[discord.Message]):
     __slots__ = "sendable", "args", "kwargs", "message"
+    sendable: discord.abc.Messageable
+    args: Any
+    kwargs: Any
+    message: Optional[discord.Message]
 
-    def __init__(self, sendable, *args, **kwargs):
+    def __init__(self, sendable: discord.abc.Messageable,
+        *args: Any, **kwargs: Any):
         self.sendable = sendable
         self.args = args
         self.kwargs = kwargs
 
-    async def __aenter__(self):
+    async def __aenter__(self) -> discord.Message:
         self.message = await self.sendable.send(*self.args, **self.kwargs)
         return self.message
 
-    async def __aexit__(self, exc_type, exc_val, tb):
+    async def __aexit__(self, exc_type, exc_val, tb) -> None: # type: ignore
         try:
-            await self.message.delete()
+            if self.message is not None:
+                await self.message.delete()
         except (discord.Forbidden, discord.NotFound):
             pass
 
 class ChannelById(discord.abc.Messageable):
     __slots__ = "id", "_state"
+    id: int
+    _state: discord.state.ConnectionState
 
-    def __init__(self, client, id):
+    def __init__(self, client: discord.Client, id: int):
         self.id = id
-        self._state = client._connection
+        self._state = client._connection # type: ignore
 
-    async def _get_channel(self):
+    async def _get_channel(self) -> discord.abc.Messageable:
         return self
-
-class MessageById:
-    __slots__ = "id"
-
-    def __init__(self, id):
-        self.id = id
-
-class RoleById:
-    __slots__ = "id"
-
-    def __init__(self, id):
-        self.id = id
